Reorganize into smaller components
